api/requirements.txt
Camelot==12.6.29
chardet==5.0.0
click==8.1.3
decorator==5.1.1
distro==1.7.0
Elixir==0.7.1
Flask==2.2.2
icalendar==4.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.23.3
pandas==1.4.4
pbr==5.10.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
six==1.16.0
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
tabula-py==2.5.1
tabulate==0.8.10
Tempita==0.5.2
Werkzeug==2.2.2
xlrd==0.7.1
xlwt==0.7.2
